Metrics: Move TimeSeriesCumulative and TimeSeriesGauge to top level. Hide TimeSeries.
